BRI-SSVEPMain-CCA-BETA/bri_requirements.txt
appdirs==1.4.4
av==10.0.0
cfclient==2023.6
cflib==0.1.23
contourpy==1.1.1
cycler==0.11.0
djitellopy==2.5.0
fonttools==4.42.1
freetype-py==2.4.0
hsluv==5.0.4
importlib-resources==6.0.1
joblib==1.3.2
kiwisolver==1.4.5
libusb-package==1.0.26.2
matplotlib==3.7.1
numpy==1.24.4
opencv-python==4.8.0.76
packaging==23.1
Pillow==10.0.1
playsound==1.3.0
pylsl==1.16.1
pyparsing==3.1.1
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
pyqtgraph==0.13.3
PySDL2==0.9.16
pysdl2-dll==2.24.0
pyserial==3.5
python-dateutil==2.8.2
pyusb==1.2.1
PyYAML==5.4.1
pyzmq==25.1.1
qasync==0.23.0
qtm==2.1.2
scikit-learn==1.3.0
scipy==1.11.2
six==1.16.0
sklearn==0.0
threadpoolctl==3.2.0
vispy==0.9.6
